services/auth-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.1
psycopg2-binary==2.9.9
pydantic==2.6.1
python-multipart==0.0.9
python-dotenv==1.0.0
